app/backend/requirements.txt
azure-cosmos
azure-identity
azure-storage-blob
openai
Flask
flask_cors
gunicorn
pandas
python-dotenv
#logging